+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Backup strategy. When one provider fail, fallback to another.
 
-## [1.0.0] - 2022-XX-YY
+## [1.1.0] - 2022-11-04
 ### Added
-- First release.
+- `ExportToStruct` renamed to `Dump`
+- `Dump` now:
+    - Set values from environment variables using the `env` field tag.
+    - Set default values using the `default` field tag.
+    - Validating the values using the `validate` field tag.